Separate tests and fix create_index
While creating tests for create_index I found a bug on
MultipleIndexesQueryTestCase. I'm fixing it in this commit.
